python_slave/requirements.txt
pyserial==3.5
requests==2.31.0
websockets==11.0.3
python-dotenv==1.0.0
asyncio==3.4.3